Alterado o tipo de Id de Long para Integer

+++ src/main/java/br/com/meli/desafiospring/models/User.java
@@ -1,7 +1,6 @@
 package br.com.meli.desafiospring.models;
 
 import br.com.meli.desafiospring.enums.UserType;
-import net.bytebuddy.implementation.bind.annotation.Default;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
@@ -15,7 +14,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @NotNull
     @Size(max = 50)
@@ -35,11 +34,11 @@ public class User {
         this.username = username;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
